Flyme :  Corner and Statusbar Height Fix
Modified Flyme-res for removing corners and other two overlays for removing the side padding of statusbar icons and setting the height to 24dp. Statusbar going to look ugly if we set height more than 24dp bcz the alignment of clock and the signal icons will be messed up. Just flymeeeeeeeeeeeeee things. kewl.